+++ src/entity/Base.ts
@@ -6,15 +6,15 @@ import {
 } from "typeorm"
 
 export abstract class Base {
-  @PrimaryGeneratedColumn()
-  id: number
+  @PrimaryGeneratedColumn("uuid")
+  id: string
 
-  @CreateDateColumn()
+  @CreateDateColumn({ type: "timestamptz" })
   created: Date
 
-  @UpdateDateColumn({ nullable: true })
+  @UpdateDateColumn({ nullable: true, type: "timestamptz" })
   updated: Date
 
-  @Column({ nullable: true })
+  @Column({ nullable: true, type: "timestamptz" })
   deleted: Date
 }